+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.2] - 2023-02-08
+### Added
+- Duration methods `daysBetween`, `monthsBetween` and `yearsBetween` added to `Dates`.
+- `format` methods with and without default formatting pattern added to `Dates`.
+
 ## [0.5.1] - 2023-02-07
 ### Added
 - Polymorphic `LocalDate` comparison functions added to utility class `Dates`.